Refactor: Split a middleware_builder into smaller submodules
