+++ axum-login/src/lib.rs
@@ -383,7 +383,6 @@ pub use axum;
 pub use backend::{AuthUser, AuthnBackend, AuthzBackend, UserId};
 #[doc(hidden)]
 pub use middleware::url_with_redirect_query;
-pub use middleware_builder::*;
 pub use service::{AuthManager, AuthManagerLayer, AuthManagerLayerBuilder};
 pub use session::{AuthSession, Error};
 pub use tower_sessions;
@@ -392,6 +391,7 @@ pub use tracing;
 mod backend;
 mod extract;
 mod middleware;
-mod middleware_builder;
 mod service;
 mod session;
+
+pub mod require;
